removed logic from create initial added helper functions and imported
